requirements.txt
# AI Agent 项目依赖
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.5.0